Rename package ui.dashboard to ui.recentactivity
