+++ app/src/main/java/com/mmutert/trackmydebt/ui/recentactivity/RecentActivityViewModel.kt
@@ -1,4 +1,4 @@
-package com.mmutert.trackmydebt.ui.dashboard
+package com.mmutert.trackmydebt.ui.recentactivity
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
