+++ app/src/main/res/layout/fragment_recent_activity.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.dashboard.RecentActivityFragment">
+        tools:context=".ui.recentactivity.RecentActivityFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:layout_width="match_parent"
